+++ src/misc.js
@@ -293,6 +293,11 @@ const singletonEvalByType = {
       return d;
     }
   },
+  "StringOrNumber": function(d){
+    if (typeof d === "string" || typeof d === "number") {
+      return d;
+    }
+  },
   "AnySingletonAtRoot": function (d) {
     return d;
   }
@@ -330,43 +335,29 @@ engine.singleton = function (coll, type) {
   throw new Error('Not supported type ' + type);
 };
 
-/**
- * Checks whether a primitve value is present
- */
-const primitives = new Set();
-// IE11 probably doesn't support `new Set(iterable)`
-[
-  "instant",
-  "time",
-  "date",
-  "dateTime",
-  "base64Binary",
-  "decimal",
-  "integer64",
-  "boolean",
-  "string",
-  "code",
-  "markdown",
-  "id",
-  "integer",
-  "unsignedInt",
-  "positiveInt",
-  "uri",
-  "oid",
-  "uuid",
-  "canonical",
-  "url",
-  "Integer",
-  "Decimal",
-  "String",
-  "Date",
-  "DateTime",
-  "Time"
-].forEach(i => primitives.add(i));
-
 engine.hasValueFn = function(coll) {
   return coll.length === 1 && util.valData(coll[0]) != null
-    && primitives.has(TypeInfo.fromValue(coll[0]).name);
+    && TypeInfo.isPrimitiveValue(coll[0]);
+};
+
+/**
+ * Returns the underlying system value for the FHIR primitive if the input
+ * collection contains a single value which is a FHIR primitive, and it has a
+ * primitive value. Otherwise, the return value is empty (i.e. []).
+ *
+ * See: https://hl7.org/fhir/fhirpath.html#functions
+ * @param {Array<*>} coll - input collection
+ * @returns {*|[]}
+ */
+engine.getValueFn = function(coll) {
+  if (coll.length === 1) {
+    const node = coll[0];
+    const v = util.valData(node);
+    if (v != null && TypeInfo.isPrimitiveValue(node)) {
+      return v;
+    }
+  }
+  return [];
 };
 
 module.exports = engine;